+++ android/app/src/main/java/com/luno/MainActivity.java
@@ -3,7 +3,6 @@ package com.luno;
 import com.facebook.react.ReactActivity;
 import com.facebook.react.ReactPackage;
 import com.facebook.react.shell.MainReactPackage;
-import com.aakashns.reactnativedialogs.ReactNativeDialogsPackage;
 
 import java.util.Arrays;
 import java.util.List;
@@ -35,7 +34,6 @@ public class MainActivity extends ReactActivity {
     @Override
     protected List<ReactPackage> getPackages() {
       return Arrays.<ReactPackage>asList(
-        new MainReactPackage(),
-        new ReactNativeDialogsPackage(this));
+        new MainReactPackage());
     }
 }