+++ src/luno/schema.cljs
@@ -3,14 +3,14 @@
 
 ;; schema of app-db
 (def schema {:android  {:drawer s/Any}
-             :ios      {:tab s/Str}
+             :shared   {:tab s/Str}
              :data     s/Any
              :error    s/Any
              :loading? s/Bool})
 
 ;; initial state of app-db
 (def app-db {:android  {:drawer nil}
-             :ios      {:tab "main"}
+             :shared   {:tab "main"}
              :data     nil
              :error    nil
              :loading? false})